fix: ensure REPLICA IDENTITY set to 'FULL'
Without this, DELETE operations cannot be broadcast to clients because they will not, in general, contain the RLS column in the change event, and we cannot determine who to send them to.

This is suboptimal because it makes replication less efficient. We can hopefully improve on this later, when we do replication into a temporary Postgres table.